Rework the implementation of CannelloniCanBackend to use Qt network stuff
Qt already has the facilities needed to run a UDP server integrated with its
QUdpSocket. Use that class in order to both send and receive Cannelloni packets
to/from the actual server.
Sending packets is rate-limited to 300 kbps, as suggested by the Cannelloni
manual. This should really be configurable at runtime in order to adapt to the
bandwidth available in the underlying physical interface. If/When such
functionality is needed, all the facilities (i.e. parametrisation) are already
there.