Bump cannelloni to v1.1.0
That enables us to start implementing support of the TCP transport,
which was recently added in cannelloni, and to reuse its decoder
of the CAN frames transmitted over TCP data streams (see files
decoder.h and decoder.cpp added in v1.1.0).

The existing code and logic of qtCannelloniCanBus are not affected
by this bump of the cannelloni submodule version.

Signed-off-by: Vladimir Beloborodov <VladimirTechMan@gmail.com>