Prevent CANdevStudio crashes on bad interface parameter strings
The current implementation of qtCannelloniCanBus assumes that a
valid pointer to a QString object was passed for the errorMessage
parameter. However, if we look at QT's documentation for the
createDevice method, that parameter can be nullptr (and it defaults
to that value if not specified). CANdevStudio does not pass that
parameter now, thus the plug-in code attempts to derefence a null
pointer when an error in the interface string is detected, which
causes CANdevStudio to crash.

This commit fixes the code so that it sets error messages only
if the error message string pointer is non-null.

Signed-off-by: Vladimir Beloborodov <VladimirTechMan@gmail.com>